Move the Android filter into its own class
