+++ webapp/src/main/java/com/box/l10n/mojito/okapi/filters/XMLFilter.java
@@ -1,25 +1,17 @@
 package com.box.l10n.mojito.okapi.filters;
 
-import com.box.l10n.mojito.service.assetExtraction.extractor.AssetPathToFilterConfigMapper;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import net.sf.okapi.common.BOMNewlineEncodingDetector;
 import net.sf.okapi.common.DefaultEntityResolver;
-import net.sf.okapi.common.Event;
-import net.sf.okapi.common.EventType;
 import net.sf.okapi.common.encoder.EncoderManager;
 import net.sf.okapi.common.exceptions.OkapiIOException;
 import net.sf.okapi.common.filters.FilterConfiguration;
-import net.sf.okapi.common.resource.Property;
 import net.sf.okapi.common.resource.StartDocument;
-import net.sf.okapi.common.resource.TextContainer;
-import net.sf.okapi.common.resource.TextUnit;
 import net.sf.okapi.common.skeleton.GenericSkeleton;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,13 +30,8 @@ public class XMLFilter extends net.sf.okapi.filters.xml.XMLFilter {
     static Logger logger = LoggerFactory.getLogger(XMLFilter.class);
 
     public static final String FILTER_CONFIG_ID = "okf_xml@mojito";
-    public static final String ANDROIDSTRINGS_CONFIG_FILE_NAME = "AndroidStrings_mojito.fprm";
     public static final String RESX_CONFIG_FILE_NAME = "resx_mojito.fprm";
     public static final String XTB_CONFIG_FILE_NAME = "xtb_mojito.fprm";
-    
-    private static final String XML_COMMENT_PATTERN = "<!--(?<comment>.*?)-->";
-    private static final String XML_COMMENT_GROUP_NAME = "comment";
-    
 
     @Override
     public String getName() {
@@ -52,7 +39,7 @@ public class XMLFilter extends net.sf.okapi.filters.xml.XMLFilter {
     }
 
     /**
-     * Overriding to include only resx and AndroidStrings filters
+     * Overriding to include only resx, xtb and AndroidStrings filters
      *
      * @return
      */
@@ -66,12 +53,6 @@ public class XMLFilter extends net.sf.okapi.filters.xml.XMLFilter {
                 "Configuration for Microsoft RESX documents (without binary data).",
                 RESX_CONFIG_FILE_NAME,
                 ".resx;"));
-        list.add(new FilterConfiguration(getName() + "-AndroidStrings",
-                getMimeType(),
-                getClass().getName(),
-                "Android Strings",
-                "Configuration for Android Strings XML documents.",
-                ANDROIDSTRINGS_CONFIG_FILE_NAME));
         list.add(new FilterConfiguration(getName() + "-xtb",
                 getMimeType(),
                 getClass().getName(),
@@ -186,92 +167,16 @@ public class XMLFilter extends net.sf.okapi.filters.xml.XMLFilter {
     }
 
     @Override
-    public Event next() {
-        Event event = super.next();
-
-        if (input.getFilterConfigId().equals(AssetPathToFilterConfigMapper.ANDROIDSTRINGS_FILTER_CONFIG_ID)
-                && event.getEventType() == EventType.TEXT_UNIT) {
-            // if source has escaped double-quotes, single-quotes, \r or \n, unescape
-            TextUnit textUnit = (TextUnit) event.getTextUnit();
-            String sourceString = textUnit.getSource().toString();
-            String unescapedSourceString = unescape(sourceString);
-            TextContainer source = new TextContainer(unescapedSourceString);
-            textUnit.setSource(source);
-            extractNoteFromXMLCommentInSkeletonIfNone(textUnit);
-        }
-
-        return event;
-    }
-
-    /**
-     * Extract the note from XML comments only if there is no note on the 
-     * text unit. In other words if a note was specify via attribute like 
-     * description for android it won't be overridden by an comments present in 
-     * the XML file.
-     * 
-     * @param textUnit the text unit for which comments should be extracted
-     */
-    protected void extractNoteFromXMLCommentInSkeletonIfNone(TextUnit textUnit) {
-
-        String skeleton = textUnit.getSkeleton().toString();
-
-        if (textUnit.getProperty(Property.NOTE) == null) {
-            String note = getNoteFromXMLCommentsInSkeleton(skeleton);
-            if (note != null) {
-                textUnit.setProperty(new Property(Property.NOTE, note));
-            }
-        }
-    }
-
-    /**
-     * Gets the note from the XML comments in the skeleton.
-     * 
-     * @param skeleton that may contains comments
-     * @return the note or <code>null</code>
-     */
-    protected String getNoteFromXMLCommentsInSkeleton(String skeleton) {
-
-        String note = null;
-
-        StringBuilder commentBuilder = new StringBuilder();
-
-        Pattern pattern = Pattern.compile(XML_COMMENT_PATTERN);
-        Matcher matcher = pattern.matcher(skeleton);
-
-        while (matcher.find()) {
-            if(commentBuilder.length() > 0) {
-                commentBuilder.append(" ");
-            }
-            commentBuilder.append(matcher.group(XML_COMMENT_GROUP_NAME).trim());
-        }
-
-        if (commentBuilder.length() > 0) {
-            note = commentBuilder.toString();
-        }
-
-        return note;
-    }
-    
-    private String unescape(String text) {
-        String unescapedText = text.replaceAll("(\\\\)(\"|')", "$2");
-        unescapedText = unescapedText.replaceAll("\\\\n", "\n");
-        unescapedText = unescapedText.replaceAll("\\\\r", "\r");
-        return unescapedText;
-    }
-
-    @Override
     public EncoderManager getEncoderManager() {
         if (encoderManager == null) {
             encoderManager = new EncoderManager();
         }
 
-        XMLEncoder xmlEncoder = new XMLEncoder();
-
-        if (input.getFilterConfigId().equals(getName() + "-AndroidStrings")) {
-            xmlEncoder.setAndroidStrings(true);
-        }
-
-        encoderManager.setMapping(getMimeType(), xmlEncoder);
+        encoderManager.setMapping(getMimeType(), getXMLEncoder());
         return encoderManager;
+    }
+
+    public XMLEncoder getXMLEncoder() {
+        return new XMLEncoder();
     }
 }
